% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denaltitude.R
\name{denaltitude}
\alias{denaltitude}
\title{Density Altitude}
\usage{
denaltitude(airportCode, elevation = "ft")
}
\arguments{
\item{airportCode}{Valid IATA airport code (character string).}

\item{elevation}{Units for elevation displayed in 'ft' (default) or 'm'.}
}
\description{
Calculate density altitude based on METAR data for a specific airport.
}
\examples{
denaltitude("FWA","ft")

}
